Time: 21 ms (100%), Space: 58.4 MB (18.75%) - LeetHub

+++ 1335-maximum-candies-allocated-to-k-children/1335-maximum-candies-allocated-to-k-children.cs
@@ -2,12 +2,12 @@ public class Solution {
     public int MaximumCandies(int[] candies, long k) {
         
         long totel = candies.Sum(c => (long)c);
-          long left = 1;
-        long right = Math.Min(totel/k , candies.Max());
+        int left = 1;
+        int right = (int)Math.Min(totel/k , candies.Max());
         if(totel>=k){
           
          while(left<=right){
-            long mid = left + (right - left)/2;
+            int mid = (left + (right - left)/2);
             long count = 0;
             foreach(int i in candies){
                 count += i/mid;
